TextProjectParser: look-up for files included as linked folders
In order to improve finding of files outside the workspace, a user may have
integrated linked folders that contain the files.

So before starting the look-up with kpsewhich, check whether the file
exists somewhere else in the project tree.

A match is only detected, when the path of the file matches the path in the
project.